+++ src/main/java/com/ycloud/client/model/WhatsappPhoneNumberProfile.java
@@ -96,7 +96,7 @@ public class WhatsappPhoneNumberProfile {
 
   public static final String SERIALIZED_NAME_WEBSITES = "websites";
   /**
-   * The URLs associated with the business. For instance, a website, Facebook Page, or Instagram. You must include the http:// or https:// portion of the URL. There is a maximum of 2 websites with a maximum of 256 characters each.
+   * The URLs associated with the business. For instance, a website, Facebook Page, or Instagram. You must include the http:// or https:// portion of the URL. There is a maximum of 2 websites with a maximum of 255 characters each.
   **/
   @SerializedName(SERIALIZED_NAME_WEBSITES)
   private List<String> websites = null;
@@ -274,7 +274,7 @@ public class WhatsappPhoneNumberProfile {
   }
 
    /**
-   * The URLs associated with the business. For instance, a website, Facebook Page, or Instagram. You must include the http:// or https:// portion of the URL. There is a maximum of 2 websites with a maximum of 256 characters each.
+   * The URLs associated with the business. For instance, a website, Facebook Page, or Instagram. You must include the http:// or https:// portion of the URL. There is a maximum of 2 websites with a maximum of 255 characters each.
    * @return websites
   **/
   @javax.annotation.Nullable
@@ -285,9 +285,9 @@ public class WhatsappPhoneNumberProfile {
 
 
   /**
-   * The URLs associated with the business. For instance, a website, Facebook Page, or Instagram. You must include the http:// or https:// portion of the URL. There is a maximum of 2 websites with a maximum of 256 characters each.
+   * The URLs associated with the business. For instance, a website, Facebook Page, or Instagram. You must include the http:// or https:// portion of the URL. There is a maximum of 2 websites with a maximum of 255 characters each.
    **/
-  @ApiModelProperty(value = "The URLs associated with the business. For instance, a website, Facebook Page, or Instagram. You must include the http:// or https:// portion of the URL. There is a maximum of 2 websites with a maximum of 256 characters each.")
+  @ApiModelProperty(value = "The URLs associated with the business. For instance, a website, Facebook Page, or Instagram. You must include the http:// or https:// portion of the URL. There is a maximum of 2 websites with a maximum of 255 characters each.")
   public void setWebsites(List<String> websites) {
     this.websites = websites;
   }